+++ core/src/main/scala/org/evactor/storage/KpiStorage.scala
@@ -19,6 +19,6 @@ import org.evactor.model.events.Event
 
 trait KpiStorage {
 
-  def getSumStatistics(channel: String, category: Option[String], fromTimestamp: Option[Long], toTimestamp: Option[Long], interval: String): (Long, List[(Long, Double)])
+  def getSumStatistics(channel: String, category: Option[String], fromTimestamp: Option[Long], toTimestamp: Option[Long], interval: String): (Long, List[(Long, Long)])
 
 }
